Google C++ Style: Classes::Access Control
Changed code to move all class member data to private access.

Had to make a shit ton of getters for members that were previously
public.

This especially sucked with vectors, which need to be passed
by reference in their getters.  There was a call higher up in the
call stack where we called AddToVector and RemoveFromVector
from the City class, and it took forever for me to realize
I needed to change that signature to reflect that the vector
is now being passed as a reference.

https://google.github.io/styleguide/cppguide.html#Access_Control

Signed-off-by: Tim Mason <masonts@gmail.com>